requirements.txt
anyio==4.0.0
Brotli==1.1.0
certifi==2023.7.22
charset-normalizer==3.3.2
crafty-client==2.0.0
h11==0.14.0
httpcore==1.0.1
httpx==0.25.1
idna==3.4
mutagen==1.47.0
pycryptodomex==3.19.0
python-telegram-bot==20.6
requests==2.31.0
sniffio==1.3.0
urllib3==2.0.7
websockets==12.0
yt-dlp==2023.10.13